+++ src/command_modules/azure-cli-appservice/azure/cli/command_modules/appservice/commands.py
@@ -144,6 +144,7 @@ cli_command(__name__, 'webapp list-runtimes', custom_path + 'list_runtimes')
 cli_command(__name__, 'webapp auth show', custom_path + 'get_auth_settings')
 cli_command(__name__, 'webapp auth update', custom_path + 'update_auth_settings')
 
+cli_command(__name__, 'webapp assign-identity', custom_path + 'assign_identity')
 
 if not supported_api_version(PROFILE_TYPE, max_api='2017-03-09-profile'):
     cli_command(__name__, 'appservice plan create', custom_path + 'create_app_service_plan', exception_handler=ex_handler_factory(creating_plan=True))
@@ -154,8 +155,7 @@ if not supported_api_version(PROFILE_TYPE, max_api='2017-03-09-profile'):
                                'azure.mgmt.web.operations.app_service_plans_operations#AppServicePlansOperations.create_or_update',
                                custom_function_op=custom_path + 'update_app_service_plan',
                                setter_arg_name='app_service_plan', factory=cf_plans)
-    cli_command(__name__, 'appservice list-locations', 'azure.mgmt.web.web_site_management_client#WebSiteManagementClient.list_geo_regions', cf_web_client, transform=transform_list_location_output)
-
+    cli_command(__name__, 'appservice list-locations', custom_path + 'list_locations', transform=transform_list_location_output)
     cli_command(__name__, 'functionapp create', custom_path + 'create_function')
     cli_command(__name__, 'functionapp list', custom_path + 'list_function_app', table_transformer=transform_web_list_output)
     cli_command(__name__, 'functionapp show', custom_path + 'show_webapp', exception_handler=empty_on_404, table_transformer=transform_web_output)
@@ -187,3 +187,4 @@ if not supported_api_version(PROFILE_TYPE, max_api='2017-03-09-profile'):
     cli_command(__name__, 'functionapp deployment list-publishing-profiles',
                 custom_path + 'list_publish_profiles')
     cli_command(__name__, 'functionapp deployment user show', 'azure.mgmt.web.web_site_management_client#WebSiteManagementClient.get_publishing_user', cf_web_client, exception_handler=empty_on_404)
+    cli_command(__name__, 'functionapp assign-identity', custom_path + 'assign_identity')