+++ src/command_modules/azure-cli-iot/azure/cli/command_modules/iot/_help.py
@@ -4,6 +4,7 @@
 # --------------------------------------------------------------------------------------------
 
 from azure.cli.core.help_files import helps
+from ._constants import certificate_help
 
 helps['iot'] = """
     type: group
@@ -20,10 +21,97 @@ helps['iot hub'] = """
     short-summary: (PREVIEW) Manage Azure IoT hubs.
 """
 
+helps['iot hub certificate'] = """
+    type: group
+    short-summary: Manage IoT Hub certificates.
+"""
+
+helps['iot hub certificate create'] = """
+    type: command
+    short-summary: Create/upload an Azure IoT Hub certificate.
+    long-summary: {0}
+    examples:
+        - name: Uploads a CA certificate PEM file to an IoT hub.
+          text: >
+            az iot hub certificate create --hub-name MyIotHub --name MyCertificate --path /certificates/Certificate.pem
+        - name: Uploads a CA certificate CER file to an IoT hub.
+          text: >
+            az iot hub certificate create --hub-name MyIotHub --name MyCertificate --path /certificates/Certificate.cer
+""".format(certificate_help)
+
+helps['iot hub certificate update'] = """
+    type: command
+    short-summary: Update an Azure IoT Hub certificate.
+    long-summary: Uploads a new certificate to replace the existing certificate with the same name. {0}
+    examples:
+        - name: Updates a CA certificate in an IoT hub by uploading a new PEM file.
+          text: >
+            az iot hub certificate update --hub-name MyIotHub --name MyCertificate --path /certificates/NewCertificate.pem --etag
+            AAAAAAAAAAA=
+        - name: Updates a CA certificate in an IoT hub by uploading a new CER file.
+          text: >
+            az iot hub certificate update --hub-name MyIotHub --name MyCertificate --path /certificates/NewCertificate.cer --etag
+            AAAAAAAAAAA=
+""".format(certificate_help)
+
+helps['iot hub certificate delete'] = """
+    type: command
+    short-summary: Deletes an Azure IoT Hub certificate.
+    long-summary: {0}
+    examples:
+        - name: Deletes MyCertificate
+          text: >
+            az iot hub certificate delete --hub-name MyIotHub --name MyCertificate --etag AAAAAAAAAAA=
+""".format(certificate_help)
+
+helps['iot hub certificate show'] = """
+    type: command
+    short-summary: Shows information about a particular Azure IoT Hub certificate.
+    long-summary: {0}
+    examples:
+        - name: Show details about MyCertificate
+          text: >
+            az iot hub certificate show --hub-name MyIotHub --name MyCertificate
+""".format(certificate_help)
+
+helps['iot hub certificate list'] = """
+    type: command
+    short-summary: Lists all certificates contained within an Azure IoT Hub
+    long-summary: {0}
+    examples:
+        - name: List all certificates in MyIotHub
+          text: >
+            az iot hub certificate list --hub-name MyIotHub
+""".format(certificate_help)
+
+helps['iot hub certificate generate-verification-code'] = """
+    type: command
+    short-summary: Generates a verification code for an Azure IoT Hub certificate.
+    long-summary: This verification code is used to complete the proof of possession step for a certificate. Use this
+                  verification code as the CN of a new certificate signed with the root certificates private key. {0}
+    examples:
+        - name: Generates a verification code for MyCertificate
+          text: >
+            az iot hub certificate generate-verification-code --hub-name MyIotHub --name MyCertificate --etag
+            AAAAAAAAAAA=
+""".format(certificate_help)
+
+helps['iot hub certificate verify'] = """
+    type: command
+    short-summary: Verifies an Azure IoT Hub certificate.
+    long-summary: Verifies a certificate by uploading a verification certificate containing the verification code obtained
+                  by calling generate-verification-code. This is the last step in the proof of possession process. {0}
+    examples:
+        - name: Verifies ownership of the MyCertificate private key.
+          text: >
+            az iot hub certificate verify --hub-name MyIotHub --name MyCertificate --path /certificates/Verification.pem --etag
+            AAAAAAAAAAA=
+""".format(certificate_help)
+
 helps['iot hub create'] = """
     type: command
     short-summary: Create an Azure IoT hub.
-    long-summary: 'For an introduction to Azure IoT Hub, see https://docs.microsoft.com/azure/iot-hub/'
+    long-summary: For an introduction to Azure IoT Hub, see https://docs.microsoft.com/azure/iot-hub/
     examples:
         - name: Create an IoT Hub with the free pricing tier F1, in the region of the resource group.
           text: >
@@ -288,11 +376,11 @@ helps['iot device message abandon'] = """
 helps['iot device export'] = """
     type: command
     short-summary: Export all the device identities in the IoT hub identity registry to an Azure Storage blob container.
-    long-summary: 'For more information, see https://docs.microsoft.com/azure/iot-hub/iot-hub-devguide-identity-registry#import-and-export-device-identities'
+    long-summary: For more information, see https://docs.microsoft.com/azure/iot-hub/iot-hub-devguide-identity-registry#import-and-export-device-identities
 """
 
 helps['iot device import'] = """
     type: command
     short-summary: Import, update, or delete device identities in the IoT hub identity registry from a blob.
-    long-summary: 'For more information, see https://docs.microsoft.com/azure/iot-hub/iot-hub-devguide-identity-registry#import-and-export-device-identities'
+    long-summary: For more information, see https://docs.microsoft.com/azure/iot-hub/iot-hub-devguide-identity-registry#import-and-export-device-identities
 """