+++ src/command_modules/azure-cli-vm/azure/cli/command_modules/vm/custom.py
@@ -2184,16 +2184,17 @@ def remove_vm_secret(resource_group_name, vm_name, keyvault, certificate=None):
 def assign_vm_identity(resource_group_name, vm_name, identity_role=DefaultStr('Contributor'),
                        identity_role_id=None, identity_scope=None, port=None):
     VirtualMachineIdentity = get_sdk(ResourceType.MGMT_COMPUTE, 'VirtualMachineIdentity', mod='models')
-    vm = get_vm(resource_group_name, vm_name)
-    if not vm.identity:
-        logger.info('Enabling managed identity...')
-        vm.identity = VirtualMachineIdentity(type='systemAssigned')
-        vm = set_vm(vm)
-    else:
-        logger.info('Managed identity is already enabled')
+    from azure.cli.core.commands.arm import assign_implict_identity
+    client = _compute_client_factory()
 
-    if identity_scope:
-        _create_role_assignment_with_retries(identity_scope, identity_role_id, vm.identity.principal_id)
+    def getter():
+        return client.virtual_machines.get(resource_group_name, vm_name)
+
+    def setter(vm):
+        vm.identity = VirtualMachineIdentity(type='SystemAssigned')
+        return set_vm(vm)
+
+    vm = assign_implict_identity(getter, setter, identity_role=identity_role_id, identity_scope=identity_scope)
 
     port = port or _MSI_PORT
     ext_name = 'ManagedIdentityExtensionFor' + ('Linux' if _is_linux_vm(vm) else 'Windows')
@@ -2209,22 +2210,20 @@ def assign_vm_identity(resource_group_name, vm_name, identity_role=DefaultStr('C
 
 def assign_vmss_identity(resource_group_name, vmss_name, identity_role=DefaultStr('Contributor'),
                          identity_role_id=None, identity_scope=None, port=None):
-    VirtualMachineIdentity, UpgradeMode = get_sdk(ResourceType.MGMT_COMPUTE, 'VirtualMachineScaleSetIdentity',
-                                                  'UpgradeMode', mod='models')
+    VirtualMachineScaleSetIdentity, UpgradeMode = get_sdk(ResourceType.MGMT_COMPUTE, 'VirtualMachineScaleSetIdentity',
+                                                          'UpgradeMode', mod='models')
+    from azure.cli.core.commands.arm import assign_implict_identity
     client = _compute_client_factory()
-    vmss = client.virtual_machine_scale_sets.get(resource_group_name, vmss_name)
-    if not vmss.identity:
-        logger.info('Enabling managed identity...')
-        vmss.identity = VirtualMachineIdentity(type='systemAssigned')
-        client.virtual_machine_scale_sets.create_or_update(resource_group_name, vmss_name, vmss)
-        # the 'create_or_update' doesn't deserialize the result right, hence we dig it out ourselves
-        # (TODO open auto-rest bug before merge)
-        vmss = client.virtual_machine_scale_sets.get(resource_group_name, vmss_name)
-    else:
-        logger.info('Managed identity is already enabled')
 
-    if identity_scope:
-        _create_role_assignment_with_retries(identity_scope, identity_role_id, vmss.identity.principal_id)
+    def getter():
+        return client.virtual_machine_scale_sets.get(resource_group_name, vmss_name)
+
+    def setter(vmss):
+        vmss.identity = VirtualMachineScaleSetIdentity(type='SystemAssigned')
+        poller = client.virtual_machine_scale_sets.create_or_update(resource_group_name, vmss_name, vmss)
+        return LongRunningOperation()(poller)
+
+    vmss = assign_implict_identity(getter, setter, identity_role=identity_role_id, identity_scope=identity_scope)
 
     port = port or _MSI_PORT
     ext_name = 'ManagedIdentityExtensionFor' + ('Linux' if vmss.virtual_machine_profile.os_profile.linux_configuration
@@ -2248,35 +2247,6 @@ def _construct_identity_info(identity_scope, identity_role, port):
         'role': str(identity_role),  # could be DefaultStr, so convert to string
         'port': port
     }
-
-
-# to workaround a known AAD server replicate issue
-def _create_role_assignment_with_retries(identity_scope, identity_role_id, principal_id):
-    import time
-    from azure.mgmt.authorization import AuthorizationManagementClient
-    from azure.mgmt.authorization.models import RoleAssignmentProperties
-    from msrestazure.azure_exceptions import CloudError
-    assignments_client = get_mgmt_service_client(AuthorizationManagementClient).role_assignments
-    properties = RoleAssignmentProperties(identity_role_id, principal_id)
-
-    logger.info("Creating an assignment with a role '%s' on the scope of '%s'", identity_role_id, identity_scope)
-    retry_times = 36
-    assignment_id = _gen_guid()
-    for l in range(0, retry_times):
-        try:
-            assignments_client.create(identity_scope, assignment_id, properties)
-            break
-        except CloudError as ex:
-            if 'role assignment already exists' in ex.message:
-                logger.info('Role assignment already exists')
-                break
-            elif l < retry_times and ' does not exist in the directory ' in ex.message:
-                time.sleep(5)
-                logger.warning('Retrying role assignment creation: %s/%s', l + 1,
-                               retry_times)
-                continue
-            else:
-                raise
 
 
 # for injecting test seams to produce predicatable role assignment id for playback